+++ CHANGELOG.md
@@ -12,9 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Output buffering mechanism to consolidate multiple output
   updates over time.
   ([#1558](https://github.com/feldera/feldera/pull/1558))
-- pipeline-manager: allow programs to be compiled in unoptimized or 
+- pipeline-manager: allow programs to be compiled in unoptimized or
   optimized modes, which tradeoff compilation speed for runtime speed
   ([#1524](https://github.com/feldera/feldera/pull/1524))
+- WebConsole: Support NULL values on Data Inspection and Insertion page (#1392)
 
 ## [0.12.0] - 2024-03-19